accessible/tests/mochitest/treeupdate/test_list_editabledoc.html

<!DOCTYPE html>
<html>

<head>
  <title>Test HTML li and listitem bullet accessible insertion into editable document</title>
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>
  <script type="application/javascript"
          src="../events.js"></script>

  <script type="application/javascript">

    // //////////////////////////////////////////////////////////////////////////
    // Invokers

    function addLi(aID) {
      this.listNode = getNode(aID);
      this.liNode = document.createElement("li");
      this.liNode.textContent = "item";

      this.eventSeq = [
        new invokerChecker(EVENT_SHOW, getAccessible, this.liNode),
        new invokerChecker(EVENT_REORDER, this.listNode)
      ];

      this.invoke = function addLi_invoke() {
        this.listNode.appendChild(this.liNode);
      };

      this.finalCheck = function addLi_finalCheck() {
        var tree = {
          role: ROLE_LIST,
          children: [
            {
              role: ROLE_LISTITEM,
              children: [
                {
                  role: ROLE_STATICTEXT,
                  name: "1. ",
                  children: []
                },
                {
                  role: ROLE_TEXT_LEAF,
                  children: []
                }
              ]
            }
          ]
        };
        testAccessibleTree(aID, tree);
      };

      this.getID = function addLi_getID() {
        return "add li";
      };
    }

    // //////////////////////////////////////////////////////////////////////////
    // Test

    // gA11yEventDumpID = "eventdump"; // debug stuff

    var gQueue = null;

    function doTest() {
      gQueue = new eventQueue();

      gQueue.push(new addLi("list"));

      gQueue.invoke(); // SimpleTest.finish() will be called in the end
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>
<body contentEditable="true">

  <a target="_blank"
     title="Wrong list bullet text of accessible for the first numbered HTML:li in CKEditor"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=557795">Mozilla Bug 557795</a>

  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <ol id="list">
  </ol>

  <div id="eventdump"></div>
</body>
</html>
